+++ R/ef_vein_eea_exhaust.R
@@ -4,6 +4,7 @@
 
 #' @name ef_vein_eea_exhaust
 #' @aliases ef_vein_eea_exhaust ef_vein_eea_exhaust_co ef_vein_eea_exhaust_nox
+#' ef_vein_eea_exhaust_ch4 ef_vein_eea_exhaust_nmhc
 #' ef_vein_eea_exhaust_pm2.5 ef_vein_eea_exhaust_pm10
 #' @description Functions to calculate vehicle speed-based hot exhaust
 #' emission factors using on EMEP/EEA guidelines 2019 methods.
@@ -13,7 +14,8 @@
 #' @param veh.segment EEA vehicle sub-classification, specific to vehicle type.
 #' For buses, embrs assigns this using veh.type and veh.wt, so it can often be
 #' ignored.
-#' @param em.type type of emission, currently CO, NOx, PM (PM2.5 or PM10; see Note).
+#' @param em.type type of emission, currently CO, NOx, CH4, NMHC and
+#' PM (PM2.5 or PM10; see Note).
 #' @param route.def route description, optional for ef_vein_eea() functions.
 #' @param eng.type engine type: ice, etc.
 #' @param eng.fuel fuel used by the vehicle: diesel, biodiesel, cng, etc.
@@ -551,7 +553,24 @@ ef_vein_eea_exhaust <-
 #ef_vein_eea_exhaust wrappers
 ###################################
 
-# eea hot exhaust nox
+
+# eea hot exhaust ch4
+#' @rdname ef_vein_eea_exhaust
+#' @export
+ef_vein_eea_exhaust_ch4 <- function(em.type = "ch4", ...){
+  #testing this
+  ef_vein_eea_exhaust(em.type = "ch4", ...)
+}
+
+# eea hot exhaust nmhc
+#' @rdname ef_vein_eea_exhaust
+#' @export
+ef_vein_eea_exhaust_nmhc <- function(em.type = "nmhc", ...){
+  #testing this
+  ef_vein_eea_exhaust(em.type = "nmhc", ...)
+}
+
+# eea hot exhaust co
 #' @rdname ef_vein_eea_exhaust
 #' @export
 ef_vein_eea_exhaust_co <- function(em.type = "co", ...){